+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `:checkhealth po` command to check the health of the plugin ([#36])
+
 ### Changed
 
 - This plugin is now loaded only when a PO file is opened ([#32])
@@ -31,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Component for [lualine.nvim](https://github.com/nvim-lualine/lualine.nvim) which prints the current translation status of the PO file ([#5])
 
+[#36]: https://github.com/toku-sa-n/po.nvim/pull/36
 [#35]: https://github.com/toku-sa-n/po.nvim/pull/35
 [#34]: https://github.com/toku-sa-n/po.nvim/pull/34
 [#32]: https://github.com/toku-sa-n/po.nvim/pull/32